am263x/am273x/awr294x: sdl: Add CCM callback feature in ESM.
Add feature in ESM to register and handle CCM callback.
Update code to fix MISRA warnings.

Fixes: PROC_SDL-5434

+++ source/sdl/esm/soc/am263x/sdl_esm_core.c
@@ -73,7 +73,7 @@ bool SDL_ESM_getBaseAddr(const SDL_ESM_Inst esmInstType, uint32_t *esmBaseAddr)
         {
             case SDL_ESM_INST_MAIN_ESM0:
                 instValid = ((bool)true);
-                *esmBaseAddr = SOC_MAIN_ESM_BASE;
+                *esmBaseAddr = (uint32_t)SOC_MAIN_ESM_BASE;
                 size = SDL_ESM0_CFG_SIZE;
                 break;
